* 'master' of git://scm.dev.nokia.troll.no/qt/qt-qa-team: (675 commits)
  Fix QDateEdit calendar pop-up having wrong date
  Repaint QProgressBar when minimum or maximum changed
  Fix QProgressBar causing timer event spam
  Fix memory leak in QCursorData
  fix gdb_dwarf_index.prf for applications, take two
  Compile fix for Mac OS X 10.7 with 10.6 sdk
  Update supported platforms and remove commercial page.
  networking: regenerate effective TLD table for cookies
  fix gdb_dwarf_index.prf for executables
  WINSCW def file update so that qtgui can pick up virtual from qtcore
  Add mobile-reduced ts-files for Symbian.
  Create cleanup stack in idleDetectorThread
  Preventing zero-timer lockups when Symbian's active scheduler used
  QNAM: Fix authentication cache when the password is in the URL
  Qt Designer: Fix static linking on Mac.
  SSL: blacklist intermediate certificates that issued weak certs
  Fix memory leak in S60Style
  Fix to QtOpenGL crash on Symbian
  Give better error message when using unsupported lookbehinds in QRegExp
  Clear error due to FRAMEBUFFER_SRGB_CAPABLE_EXT
  ...

+++ src/gui/graphicsview/qgridlayoutengine.cpp
@@ -84,7 +84,7 @@ static qreal fixedDescent(qreal descent, qreal ascent, qreal targetSize)
     Q_ASSERT(targetSize >= ascent + descent);
 
     qreal extra = targetSize - (ascent + descent);
-    return descent + (extra / 2.0);
+    return descent + (extra / qreal(2.0));
 }
 
 static qreal compare(const QGridLayoutBox &box1, const QGridLayoutBox &box2, int which)
@@ -292,9 +292,9 @@ void QGridLayoutRowData::calculateGeometries(int start, int end, qreal targetSiz
                     int stretch = stretches[start + i];
                     if (sumStretches == 0) {
                         if (hasIgnoreFlag) {
-                            factors[i] = (stretch < 0) ? 1.0 : 0.0;
+                            factors[i] = (stretch < 0) ? qreal(1.0) : qreal(0.0);
                         } else {
-                            factors[i] = (stretch < 0) ? sizes[i] : 0.0;
+                            factors[i] = (stretch < 0) ? sizes[i] : qreal(0.0);
                         }
                     } else if (stretch == sumStretches) {
                         factors[i] = 1.0;
@@ -373,7 +373,7 @@ void QGridLayoutRowData::calculateGeometries(int start, int end, qreal targetSiz
 
             for (int i = 0; i < n; ++i) {
                 if (newSizes[i] < 0.0) {
-                    qreal delta = (sumFactors == 0.0) ? 0.0
+                    qreal delta = (sumFactors == 0.0) ? qreal(0.0)
                                                       : sumCurrentAvailable * factors[i] / sumFactors;
                     newSizes[i] = sizes[i] + delta;
                 }
